many changes made to memory

+++ lib/wlib/memory/Allocator.h
@@ -6,19 +6,17 @@
  * or provided by the user. The goal is to eliminate memory fragmentation and make allocations
  * faster and much safer
  *
+ * This class should never be used because it is used internally
+ *
  * @author Deep Dhillon
- * @date October 22, 2017
+ * @date November 19, 2017
  * @bug No known bugs
  */
 
 #ifndef C_TEST_ALLOCATOR_H
 #define C_TEST_ALLOCATOR_H
 
-
-#include <stddef.h>
 #include <stdint.h>
-
-#include "../Wlib.h"
 
 namespace wlp {
     class Allocator {
@@ -57,7 +55,7 @@ namespace wlp {
          * @param blockSize size of memory blocks that can acquired at a time
          * @param poolSize size of memory pool to be created
          */
-        explicit Allocator(uint16_t blockSize, uint16_t poolSize = 0);
+        explicit Allocator(uint16_t blockSize, uint32_t poolSize = 0);
 
         /**
          * Allocator used for allocating memory where memory is provided to the allocator. It uses the given
@@ -74,7 +72,7 @@ namespace wlp {
          * @param poolSize size of the memory pool provided
          * @param type type of memory pool provided (static or dynamic)
          */
-        Allocator(uint16_t blockSize, void *pPool, uint16_t poolSize, Type type);
+        Allocator(uint16_t blockSize, void *pPool, uint32_t poolSize, Type type);
 
         /**
          * Deletes memory and returns it back to the system
@@ -123,7 +121,7 @@ namespace wlp {
          *
          * @return size of memory block
          */
-        inline size_t GetBlockSize() const {
+        inline uint16_t GetBlockSize() const {
             return m_blockSize;
         }
 
@@ -132,7 +130,7 @@ namespace wlp {
          *
          * @return size of pool
          */
-        inline size_t GetPoolSize() const {
+        inline uint32_t GetPoolSize() const {
             return m_poolSize;
         }
 
@@ -195,14 +193,14 @@ namespace wlp {
          * @param allocationType type of memory in memory pool
          * @param pPool address to memory provided
          */
-        explicit Allocator(uint16_t blockSize, uint16_t poolSize, Allocator::Type allocationType, void *pPool);
+        explicit Allocator(uint16_t blockSize, uint32_t poolSize, Allocator::Type allocationType, void *pPool);
 
 
         Type m_poolType;
-        size_t m_blockSize;
+        uint16_t m_blockSize;
         Block *m_pHead;
         Block *m_pPool;
-        size_t m_poolSize;
+        uint32_t m_poolSize;
         uint16_t m_poolTotalBlockCnt;
         uint16_t m_poolCurrBlockCnt;
         uint16_t m_totalBlockCount;
